requirements.txt
Flask==3.0.0
requests==2.31.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
gunicorn==21.2.0

